+++ src/main/java/org/sagebionetworks/bridge/spring/controllers/MasterSchedulerController.java
@@ -5,7 +5,8 @@ import java.util.List;
 import static org.sagebionetworks.bridge.Roles.ADMIN;
 import org.sagebionetworks.bridge.models.ResourceList;
 import org.sagebionetworks.bridge.models.StatusMessage;
-import org.sagebionetworks.bridge.models.TimestampHolder;
+import org.sagebionetworks.bridge.models.DateTimeHolder;
+import org.sagebionetworks.bridge.models.accounts.UserSession;
 import org.sagebionetworks.bridge.models.schedules.MasterSchedulerConfig;
 import org.sagebionetworks.bridge.services.MasterSchedulerService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,18 +23,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class MasterSchedulerController extends BaseController {
     static final StatusMessage DELETED_MSG = new StatusMessage("Scheduler config deleted.");
-
+    
+    private MasterSchedulerService schedulerService;
     
     @Autowired
-    MasterSchedulerService schedulerService;
-
-    public void setMasterSchedulerService(MasterSchedulerService schedulerService) {
+    final void setMasterSchedulerService(MasterSchedulerService schedulerService) {
         this.schedulerService = schedulerService;
     }
 
     @GetMapping("/v3/schedulerconfigs")
     public ResourceList<MasterSchedulerConfig> getAllSchedulerConfigs() {
-        getAuthenticatedSession(ADMIN);
+        UserSession session = getAuthenticatedSession(ADMIN);
+        verifyCrossStudyAdmin(session.getId(), "Study admin cannot get all scheduler configs.");
         
         List<MasterSchedulerConfig> configs = schedulerService.getAllSchedulerConfigs();
         
@@ -43,7 +44,8 @@ public class MasterSchedulerController extends BaseController {
     @PostMapping("/v3/schedulerconfigs")
     @ResponseStatus(HttpStatus.CREATED)
     public MasterSchedulerConfig createSchedulerConfig() {
-        getAuthenticatedSession(ADMIN);
+        UserSession session = getAuthenticatedSession(ADMIN);
+        verifyCrossStudyAdmin(session.getId(), "Study admin cannot create a scheduler config.");
         
         MasterSchedulerConfig schedulerConfig = parseJson(MasterSchedulerConfig.class);
         
@@ -55,7 +57,8 @@ public class MasterSchedulerController extends BaseController {
      */
     @GetMapping("/v3/schedulerconfigs/{scheduleId}")
     public MasterSchedulerConfig getSchedulerConfig(@PathVariable String scheduleId) {
-        getAuthenticatedSession(ADMIN);
+        UserSession session = getAuthenticatedSession(ADMIN);
+        verifyCrossStudyAdmin(session.getId(), "Study admin cannot get a scheduler config.");
         
         return schedulerService.getSchedulerConfig(scheduleId);
     }
@@ -65,7 +68,8 @@ public class MasterSchedulerController extends BaseController {
      */
     @PostMapping("/v3/schedulerconfigs/{scheduleId}")
     public MasterSchedulerConfig updateSchedulerConfig(@PathVariable String scheduleId) {
-        getAuthenticatedSession(ADMIN);
+        UserSession session = getAuthenticatedSession(ADMIN);
+        verifyCrossStudyAdmin(session.getId(), "Study admin cannot update a scheduler config.");
         
         MasterSchedulerConfig schedulerConfig = parseJson(MasterSchedulerConfig.class);
         
@@ -77,7 +81,8 @@ public class MasterSchedulerController extends BaseController {
      */
     @DeleteMapping("/v3/schedulerconfigs/{scheduleId}")
     public StatusMessage deleteSchedulerConfig(@PathVariable String scheduleId) {
-        getAuthenticatedSession(ADMIN);
+        UserSession session = getAuthenticatedSession(ADMIN);
+        verifyCrossStudyAdmin(session.getId(), "Study admin cannot delete a scheduler config.");
         
         schedulerService.deleteSchedulerConfig(scheduleId);
         return DELETED_MSG;
@@ -88,8 +93,9 @@ public class MasterSchedulerController extends BaseController {
      * Get the last time scheduler ran.
      */
     @GetMapping("/v3/schedulerstatus")
-    public TimestampHolder getSchedulerStatus() {
-        getAuthenticatedSession(ADMIN);
+    public DateTimeHolder getSchedulerStatus() {
+        UserSession session = getAuthenticatedSession(ADMIN);
+        verifyCrossStudyAdmin(session.getId(), "Study admin cannot get the scheduler status.");
         
         return schedulerService.getSchedulerStatus();
     }
